Splits unpackers.common into a package
