+++ reprounzip/setup.py
@@ -17,7 +17,8 @@ if sys.version_info < (2, 7):
     req.append('argparse')
 setup(name='reprounzip',
       version='0.5.1',
-      packages=['reprounzip', 'reprounzip.unpackers', 'reprounzip.plugins'],
+      packages=['reprounzip', 'reprounzip.unpackers',
+                'reprounzip.unpackers.common', 'reprounzip.plugins'],
       entry_points={
           'console_scripts': [
               'reprounzip = reprounzip.main:main'],
